implementation de tous les modules (crud) et containerisation ...

+++ src/main/java/com/immobilier/gestionImmobiliere/donnees/biens/repository/CourRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CourRepository extends JpaRepository<Cour, Integer> {
-    Page<Cour> findBySecteur_IdSecteur(Long idSecteur, Pageable pageable);
+    Page<Cour> findBySecteur_IdSecteur(Integer idSecteur, Pageable pageable);
     Page<Cour> findByProprietaire_IdUser(Integer idUser, Pageable pageable);
 }
