+++ src/main/java/com/immobilier/gestionImmobiliere/donnees/localisation/repository/PaysRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PaysRepository extends JpaRepository<Pays, Long> {
+public interface PaysRepository extends JpaRepository<Pays, Integer> {
     boolean existsByCodePaysIgnoreCase(String codePays);
     Optional<Pays> findByCodePaysIgnoreCase(String codePays);
     Page<Pays> findAll(Pageable pageable);
