+++ src/main/java/com/immobilier/gestionImmobiliere/exceptions/SecteurNotFoundException.java
@@ -1,7 +1,7 @@
 package com.immobilier.gestionImmobiliere.exceptions;
 
 public class SecteurNotFoundException extends RuntimeException {
-    public SecteurNotFoundException(Long id) {
+    public SecteurNotFoundException(Integer id) {
         super("Aucune ville trouvée avec l'id : " + id);
     }
 }
