+++ src/main/java/com/immobilier/gestionImmobiliere/exceptions/VilleNotFoundException.java
@@ -1,7 +1,7 @@
 package com.immobilier.gestionImmobiliere.exceptions;
 
 public class VilleNotFoundException extends RuntimeException {
-    public VilleNotFoundException(Long id) {
+    public VilleNotFoundException(Integer id) {
         super("Aucune ville trouvée avec l'id : " + id);
     }
 }
